+++ lib/src/generators/state/getx/getx_project_files.dart
@@ -3,10 +3,7 @@ import '../../../templates/core/localization_templates.dart';
 import '../common_project_files.dart';
 
 Map<String, String> buildGetxProjectFiles({String projectName = 'my_app'}) {
-  final files = buildCommonProjectFiles(
-    StateManagement.getx,
-    projectName: projectName,
-  );
+  final files = buildCommonProjectFiles(StateManagement.getx);
 
   files.addAll({
     'lib/core/localization/getx_localization.dart':